Add panelcontents to index.
